Python/Pyrogram/Image-Compressor-Bot/requirements.txt
pyrogram
python-dotenv
tinify
